- Performance measure added to page layout, CSS modified for performance measure w.r.t design, Background-color added for Page
